+++ Controllers/PersonCasesController.cs
@@ -20,6 +20,7 @@ namespace TrojWebApp.Controllers
         private readonly PersonTypesConnection _personTypesConnection;
         private readonly PersonsConnection _personConnection;
         private readonly UserConnection _userConnection;
+        private static int _currentCaseId;
 
         public PersonCasesController(TrojContext context, IConfiguration configuration, UserManager<IdentityUser> userManager) : base(userManager)
         {
@@ -34,10 +35,14 @@ namespace TrojWebApp.Controllers
         {
             if (id == null)
                 return NoContent();
+            _currentCaseId = id.Value;
 
             CasesViewModel currentCase = await _caseConnection.GetCase(id.Value);
             if (currentCase == null)
                 return NoContent();
+
+            var permission = _userConnection.AccessToSubPage(HttpContext.Request, UserName);
+            if (!permission) return RedirectToAction("Details", "Cases", new { id = _currentCaseId });
 
             IEnumerable<SubPageMenusChildViewModel> menu = await _userConnection.GetMenu(HttpContext.Request, UserName);
             ViewBag.Menu = menu;
@@ -65,6 +70,9 @@ namespace TrojWebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(IFormCollection collection)
         {
+            var permission = _userConnection.AccessToSubPage(HttpContext.Request, UserName);
+            if (!permission) return RedirectToAction("Details", "Cases", new { id = _currentCaseId });
+
             if (!collection.TryGetValue("CaseId", out StringValues caseId))
                 return NoContent();
 
@@ -84,6 +92,9 @@ namespace TrojWebApp.Controllers
         // GET: PersonCasesController/Delete/5
         public async Task<ActionResult> Delete(int id, int caseId)
         {
+            var permission = _userConnection.AccessToSubPage(HttpContext.Request, UserName);
+            if (!permission) return RedirectToAction("Details", "Cases", new { id = caseId });
+
             var response = await _caseConnection.DeleteCasePerson(id);
             if (response == 0)
                 return NoContent();
